==> 1.rename.ipynb <==
import os

folder_path = 'Historical_Data'

modifiedfiles = 0
for filename in os.listdir(folder_path):
    if filename.endswith('.csv') and ' ' in filename:
        stock_symbol = filename.split(' ')[0]
        new_filename = os.path.join(folder_path, stock_symbol + '.csv')
        os.rename(os.path.join(folder_path, filename), new_filename)
        modifiedfiles += 1
print(f'{modifiedfiles} files renamed.')
